Refactoring recurrent decoder
Recurrent decoder becomes a child class of an abstract SequenceDecoder,
which will be subsequently used as a base child for the TransformerDecoder.

SequenceDecoder is a model part which handles its own (train) inputs and
project the decoder outputs to logits over vocabulary and then softmax.
It also defines its cost and logits.

Beam search should be rewritten to work with this class.

For LoopState, (temporary) hack is implemented - the LoopState now contains a
wildcard placeholder for any loop state objects of the children of the
SequenceDecoder class.